datatype: fix alignment code
The alignment code in MPICH is incorrect in that it assumes that, for
basic types, the size and alignments match up.  This is fundamentally
incorrect.  On 32-bit platforms, for example, 64-bit types are
emulated, so their sizes might be 8 bytes, but their alignment
requires are still 4 bytes.  This also holds for larger types, such as
long double, where some platforms (e.g., Cygwin) define it as 12
bytes, but their alignment requirement is 16 bytes.  In short, size is
not a proxy for alignment and should not be used as such.

A side benefit of using native alignment checks is that the code is
much simpler than earlier.